// src/core/server/events/types.ts
export enum CoralEventType {
  COMMENT_ENTERED_MODERATION_QUEUE = "COMMENT_ENTERED_MODERATION_QUEUE",
  COMMENT_LEFT_MODERATION_QUEUE = "COMMENT_LEFT_MODERATION_QUEUE",
  COMMENT_STATUS_UPDATED = "COMMENT_STATUS_UPDATED",
  COMMENT_REPLY_CREATED = "COMMENT_REPLY_CREATED",
  COMMENT_CREATED = "COMMENT_CREATED",
  COMMENT_FEATURED = "COMMENT_FEATURED",
  COMMENT_RELEASED = "COMMENT_RELEASED",
  STORY_CREATED = "STORY_CREATED",
  COMMENT_REACTION_CREATED = "COMMENT_REACTION_CREATED",
  COMMENT_FLAG_CREATED = "COMMENT_FLAG_CREATED",
}